+++ src/AVPRIndex/MD5Hash.fs
@@ -29,4 +29,3 @@ type Hash =
         path
         |> File.ReadAllText
         |> Hash.hashString
-